requirements.txt
google-genai>=1.0.0
python-dotenv
streamlit
PyMuPDF==1.23.21
